+++ apps/api/src/expenses/expenses.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
-import { ExpensesService } from './expenses.service';
+import { ExpensesService } from './services/expenses.service';
 import { ExpensesController } from './expenses.controller';
+import { ExpenseStatsService } from './services/expense-stats.service';
 
 @Module({
-  providers: [ExpensesService],
+  providers: [ExpensesService, ExpenseStatsService],
   controllers: [ExpensesController],
+  exports: [ExpensesService, ExpenseStatsService],
 })
 export class ExpensesModule {}
